Got rid of nom and refactored ast into it's own module
